SZTzal/mysite/requirements.txt
appdirs==1.4.0
defusedxml==0.5.0
Django==1.10.5
django-bootstrap==0.2.4
django-bootstrap-themes==3.3.6
django-bootstrap3==8.1.0
django-forms-bootstrap==3.0.1
django-registration==2.2
django-social-auth==0.7.28
django-stdimage==2.4.1
future==0.16.0
httplib2==0.10.3
oauth2==1.9.0.post1
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
progressbar2==3.12.0
pycryptodomex==3.4.3
pyjwkest==1.3.2
PyJWT==1.4.2
pyparsing==2.1.10
python-openid==2.2.5
python-social-auth==0.2.21
python-utils==2.0.1
python3-openid==3.0.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
